замена функции atoi на strtol, изменён тип данных функции timer_from

+++ project/include/utils.h
@@ -2,7 +2,7 @@
 #define UTILS_H
 
 
-size_t timer_from(unsigned char from);
+int timer_from(unsigned char from);
 // TODO: Implement `power of` function
 // int custom_pow(int base, int power);
